requirements.txt
Flask
Gunicorn
Requests
python-dotenv
jsonify 
sentry_sdk